+++ packages/utils/src/schema/index.ts
@@ -8,8 +8,8 @@ import getFromSchema from './getFromSchema';
 import isFilesArray from './isFilesArray';
 import isMultiSelect from './isMultiSelect';
 import isSelect from './isSelect';
-import omitExtraData, { getUsedFormData, getFieldNames } from './omitExtraData';
-import retrieveSchema from './retrieveSchema';
+import omitExtraData, { getUsedFormData, getFieldNames, isValueEmpty } from './omitExtraData';
+import retrieveSchema, { relaxOptionsForScoring } from './retrieveSchema';
 import sanitizeDataForNewSchema from './sanitizeDataForNewSchema';
 import toPathSchema from './toPathSchema';
 
@@ -26,7 +26,9 @@ export {
   isFilesArray,
   isMultiSelect,
   isSelect,
+  isValueEmpty,
   omitExtraData,
+  relaxOptionsForScoring,
   retrieveSchema,
   sanitizeDataForNewSchema,
   toPathSchema,